lint cleanups

+++ src/pager.h
@@ -131,7 +131,7 @@ class pager_t {
     void set_prefix(const wcstring &pref);
 
     // Sets the terminal width and height.
-    void set_term_size(int w, int h);
+    void set_term_size(size_t w, size_t h);
 
     // Changes the selected completion in the given direction according to the layout of the given
     // rendering. Returns true if the selection changed.
